# perf: Accelerate cloth grid generation by ~7x.