## Description
Builder::add_cloth_grid() shows prohibitive runtime (>100×100 grid size), slowing down test iterations.

Performance improvements:

| Grid Size |   Before   |  After    |
|------------------------------------|
|  100x100  |   35 sec.  |   5 sec.  |
|  200x200  |  135 sec.  |  15 sec.  |


## Newton Migration Guide

Please ensure the migration guide for **warp.sim** users is up-to-date with the changes made in this MR.

- [x] The migration guide in ``docs/migration.rst`` is up-to date

## Before your PR is "Ready for review"

- [x] All commits are [signed-off](https://git-scm.com/docs/git-commit#Documentation/git-commit.txt--s) to indicate that your contribution adheres to the [Developer Certificate of Origin](https://developercertificate.org/) requirements
- [x] I understand that **GitHub** does not perform any GPU testing of this pull request
- [ ] Necessary tests have been added
- [ ] Documentation is up-to-date
- [ ] Code passes formatting and linting checks with `pre-commit run -a`