@AnkaChan 2025-05-29
They first assemble the cloth vertices and faces, then call `add_cloth_mesh` to add them as an FEM mesh. There is one caveat though: in the old `add_cloth_grid` function all the particles has the same mass but now the vertices on the boundary will have less mass (since they don't have all the neighboring faces). This might change the behavior of previous applications using `add_cloth_grid`. @eric-heiden What do you think of this? Is it okay to change the behavior?