@WenchaoHuang 2025-05-29
You're right, that's the correct approach. It's been fixed.